Added `de/helpers` module
